backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiohttp==3.9.1
paddleocr==2.7.0
paddlepaddle==2.6.0
opencv-python==4.9.0
pillow==10.2.0
apscheduler==3.10.4
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
